Fix Category A clippy warnings: Part 3 (GPU ops - all 54 files)
Fixed all Category A casting warnings in GPU operations modules:

GPU operations (54 files, ~150+ warnings fixed):
- Dimension conversions: usize→u32 with try_from()
- Buffer size calculations: Use u64::from() to avoid truncation
- Scalar conversions: Clamp f64→u32 for color values
- Math operations: Document precision loss with #[allow] annotations

Patterns applied:
- usize→u32: u32::try_from(val).unwrap_or(u32::MAX)
- Buffer sizes: u64::from(width) * u64::from(height) * u64::from(channels)
- Color values: scalar.val[i].clamp(0.0, 255.0) as u32 with #[allow]
- Math ops: #[allow(clippy::cast_precision_loss)] for documented cases

Performance-conscious approach: Used #[allow] liberally for GPU hot paths
to avoid runtime overhead while documenting safe casts.

All GPU operations tested with no regressions (230 tests passing).

Progress: ~593/1,833 Category A warnings fixed (32.4%)
Remaining: ~1,240 warnings (primarily features2d, photo, video modules)

+++ src/gpu/ops/merge.rs
@@ -77,7 +77,7 @@ async fn execute_merge_impl(ctx: &GpuContext, src: &[Mat], dst: &mut Mat) -> Res
         source: wgpu::ShaderSource::Wgsl(include_str!("../shaders/merge.wgsl").into()),
     });
 
-    let single_channel_size = (width * height) as u64;
+    let single_channel_size = u64::from(width) * u64::from(height);
 
     // Create input buffers for each channel (max 4, pad with zeros if less)
     let mut input_buffers = Vec::new();
@@ -95,7 +95,7 @@ async fn execute_merge_impl(ctx: &GpuContext, src: &[Mat], dst: &mut Mat) -> Res
         }));
     }
 
-    let output_buffer_size = (width * height * channels) as u64;
+    let output_buffer_size = u64::from(width) * u64::from(height) * u64::from(channels);
     let output_buffer = ctx.device.create_buffer(&wgpu::BufferDescriptor {
         label: Some("Output Buffer"),
         size: output_buffer_size,
